+++ Proyecto-Core/ejbModule/com/core/data/persistencia/interfaces/remotas/CatastrofeDAORemota.java
@@ -8,7 +8,7 @@ import com.core.data.entites.Catastrofe;
 @Remote
 public interface CatastrofeDAORemota {
 
-	public Integer insert(Catastrofe catastrofe)  throws Exception;
+	public Long insert(Catastrofe catastrofe)  throws Exception;
 	
 	public boolean existeCatastrofe(String nombreEvento);
