+++ Proyecto-Core/ejbModule/com/core/service/negocio/remote/CatastrofeEBR.java
@@ -15,7 +15,7 @@ import com.core.data.entites.Servicio;
 @Remote
 public interface CatastrofeEBR {
 
-	public Integer ingesarCatastrofe(String nombreEvento, String descripcion, String logo, BigDecimal coordenadasX,BigDecimal coordenadasY, Boolean activa, Boolean prioridad, Collection<Servicio> servicios, Collection<Ong> ongs,PlanDeRiesgo planDeRiesgo)throws Exception;
+	public Long ingesarCatastrofe(String nombreEvento, String descripcion, String logo, BigDecimal coordenadasX,BigDecimal coordenadasY, Boolean activa, Boolean prioridad, Collection<Servicio> servicios, Collection<Ong> ongs,PlanDeRiesgo planDeRiesgo)throws Exception;
 	//Retorna 0 si no puede ingresar la catastrofe, de lo contrario devuelve el id 
 	
 	public Catastrofe buscaCatastrofe(String nombreEvento) throws Exception;
